+++ src/css/HomePage.styles.css
@@ -125,9 +125,13 @@
     display: flex;
     overflow-y: hidden;
     overflow-x: scroll;
-    margin: 0 -25px;
-    width: calc(100% + 40px);
+    margin-left: -20px;
+    width: 100vw;
     padding: 0 20px;
+  }
+
+  .Home-podcast:last-child {
+    padding-right: 20px;
   }
 
   .Home-podcast-title {
